agents/supervisor/code/pyproject.toml
[project]
name = "supervisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "strands-agents>=1.19.0",
    "strands-agents-tools>=0.1.0",
    "uvicorn[standard]>=0.38.0",
    "hypothesis>=6.0.0",
    "bedrock-agentcore[strands-agents]>=0.1.0",
]
